LPS-116999 Moves to com.liferay.journal.internal.exportimport.content.processor
