+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/exportimport/lifecycle/JournalCacheExportImportLifecycleListener.java
@@ -19,7 +19,7 @@ import com.liferay.exportimport.kernel.lifecycle.EventAwareExportImportLifecycle
 import com.liferay.exportimport.kernel.lifecycle.ExportImportLifecycleListener;
 import com.liferay.exportimport.kernel.model.ExportImportConfiguration;
 import com.liferay.journal.constants.JournalPortletKeys;
-import com.liferay.journal.internal.exportimport.JournalArticleExportImportCache;
+import com.liferay.journal.internal.exportimport.content.processor.JournalArticleExportImportCache;
 import com.liferay.journal.util.JournalContent;
 import com.liferay.portal.kernel.model.StagedModel;
